# clerk-react/openapi.json
{"openapi":"3.1.0","info":{"title":"Questions API","version":"0.1.0"},"paths":{"/api/v1/questionsquestionlist":{"get":{"tags":["questions"],"summary":"Read Questionslist","description":"根据 user_id 查询题目列表。","operationId":"questions-read_questionslist","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"items":{"$ref":"#/components/schemas/QuestionList"},"type":"array","title":"Response Questions-Read Questionslist"}}}}}}},"/api/v1/questions/{question_id}":{"get":{"tags":["questions"],"summary":"Read Question","description":"根据 question_id 查询完整题目与答案。","operationId":"questions-read_question","parameters":[{"name":"question_id","in":"path","required":true,"schema":{"type":"string","format":"uuid","title":"Question Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/QuestionPublic"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/questions/":{"post":{"tags":["questions"],"summary":"Create Question","description":"创建题目。","operationId":"questions-create_question","parameters":[{"name":"difficulty","in":"query","required":true,"schema":{"$ref":"#/components/schemas/QuestionDifficulty"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/QuestionPublic"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/quota/quota":{"get":{"tags":["quota"],"summary":"Read Quota","description":"根据 user_id 查询配额。","operationId":"quota-read_quota","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ChallengeQuotaPublic"}}}}}}}},"components":{"schemas":{"ChallengeQuotaPublic":{"properties":{"remaining_quota":{"type":"integer","minimum":0.0,"title":"Remaining Quota","description":"剩余引用次数","default":10},"daily_quota":{"type":"integer","minimum":0.0,"title":"Daily Quota","description":"每日配额","default":10}},"type":"object","title":"ChallengeQuotaPublic","description":"挑战配额响应模型。"},"HTTPValidationError":{"properties":{"detail":{"items":{"$ref":"#/components/schemas/ValidationError"},"type":"array","title":"Detail"}},"type":"object","title":"HTTPValidationError"},"Option":{"properties":{"option_index":{"type":"integer","title":"Option Index"},"option_text":{"type":"string","title":"Option Text"}},"type":"object","required":["option_index","option_text"],"title":"Option"},"QuestionDifficulty":{"type":"string","enum":["easy","medium","hard"],"title":"QuestionDifficulty"},"QuestionList":{"properties":{"difficulty":{"$ref":"#/components/schemas/QuestionDifficulty","description":"题目难度","default":"easy"},"question_description":{"anyOf":[{"type":"string","maxLength":500},{"type":"null"}],"title":"Question Description","description":"简短描述"},"created_at":{"type":"string","format":"date-time","title":"Created At","description":"创建时间"},"id":{"type":"string","format":"uuid","title":"Id","description":"题目id"}},"type":"object","required":["question_description","id"],"title":"QuestionList","description":"题目列表模型。"},"QuestionPublic":{"properties":{"difficulty":{"$ref":"#/components/schemas/QuestionDifficulty","description":"题目难度","default":"easy"},"question_description":{"anyOf":[{"type":"string","maxLength":500},{"type":"null"}],"title":"Question Description","description":"简短描述"},"created_at":{"type":"string","format":"date-time","title":"Created At","description":"创建时间"},"id":{"type":"string","format":"uuid","title":"Id","description":"题目id"},"options":{"items":{"$ref":"#/components/schemas/Option"},"type":"array","title":"Options","description":"题目选项"},"correct_option_index":{"type":"integer","title":"Correct Option Index","description":"正确选项索引"},"question_content":{"type":"string","maxLength":1000,"title":"Question Content","description":"题目内容"},"explanation":{"type":"string","title":"Explanation","description":"解释"}},"type":"object","required":["question_description","id","options","correct_option_index","question_content","explanation"],"title":"QuestionPublic","description":"题目公开响应模型。"},"ValidationError":{"properties":{"loc":{"items":{"anyOf":[{"type":"string"},{"type":"integer"}]},"type":"array","title":"Location"},"msg":{"type":"string","title":"Message"},"type":{"type":"string","title":"Error Type"}},"type":"object","required":["loc","msg","type"],"title":"ValidationError"}}}}